Bases for ChallangesVW and DAO done

+++ code/app/src/main/java/com/example/nomoosugar/db/ChallengeEntity.kt
@@ -5,7 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "challenges")
 data class ChallengeEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0L,
+
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
     val name: String,
     val startDate: Long,
     val endDate: Long? = null,
